backend/requirements.txt
fastapi[all]==0.115.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.30.0
pgvector==0.3.5
psycopg2-binary==2.9.10
redis==5.0.8
openai==1.52.2
google-generativeai==0.8.3
sentence-transformers==3.2.1
torch==2.4.1
transformers==4.46.1
PyMuPDF==1.24.10
unstructured==0.16.11
python-multipart==0.0.12
httpx==0.27.2
orjson==3.10.7
loguru==0.7.2
tenacity==9.0.0
numpy==1.26.4
scikit-learn==1.5.2

